// connector-backend/src/__generated__/identity.ts
/**
 * !!! This file is autogenerated do not edit by hand !!!
 *
 * Generated by: @databases/pg-schema-print-types
 * Checksum: GJo83QzXIw5wz3GMjbTp/bRkb8KFcYS+oct/MHp4p3B+iahywbWFTBNxiQjKF+gKYii7T7ByOxOBJcFlenpJcA==
 */

/* eslint-disable */
// tslint:disable

interface Identity {
  did: string & {readonly __brand?: 'identity_did'}
  privkey: string
}
export default Identity;

interface Identity_InsertParameters {
  did: string & {readonly __brand?: 'identity_did'}
  privkey: string
}
export type {Identity_InsertParameters}